scoring_model/requirements.txt
uvicorn>=0.27.0
openai>=1.9.0
python-dotenv>=1.0.1
pydantic>=2.6.0
typing-extensions>=4.9.0
numpy>=1.26.0
langchain>=0.1.4
langchain-openai>=0.0.5
requests>=2.31.0 